Working on slash handling
