--- parser/tokenizerNew.js
@@ -1,351 +1,0 @@
-import { replaceTagNameWithPound } from '../utils/hedStrings'
-import { unicodeName } from 'unicode-name'
-import { generateIssue } from '../common/issues/issues'
-
-const CHARACTERS = {
-  BLANK: ' ',
-  OPENING_GROUP: '(',
-  CLOSING_GROUP: ')',
-  OPENING_COLUMN: '{',
-  CLOSING_COLUMN: '}',
-  COMMA: ',',
-  COLON: ':',
-  SLASH: '/',
-}
-
-function getTrimmedBounds(originalString) {
-  const start = originalString.search(/\S/)
-  const end = originalString.search(/\S\s*$/)
-
-  if (start === -1) {
-    // The string contains only whitespace
-    return null
-  }
-
-  return [start, end + 1]
-}
-
-const invalidCharacters = new Set(['[', ']', '~', '"'])
-// Add control codes to invalidCharacters
-for (let i = 0x00; i <= 0x1f; i++) {
-  invalidCharacters.add(String.fromCodePoint(i))
-}
-for (let i = 0x7f; i <= 0x9f; i++) {
-  invalidCharacters.add(String.fromCodePoint(i))
-}
-
-const invalidCharactersOutsideOfValues = new Set([':'])
-
-/**
- * A specification for a tokenized substring.
- */
-export class SubstringSpec {
-  /**
-   * The starting and ending bounds of the substring.
-   * @type {number[]}
-   */
-  bounds
-
-  constructor(start, end) {
-    this.bounds = [start, end]
-  }
-}
-
-/**
- * A specification for a tokenized tag.
- */
-export class TagSpec extends SubstringSpec {
-  /**
-   * The tag this spec represents.
-   * @type {string}
-   */
-  tag
-  /**
-   * The schema prefix for this tag, if any.
-   * @type {string}
-   */
-  library
-
-  constructor(tag, start, end, librarySchema) {
-    super(start, end)
-
-    this.tag = tag.trim()
-    this.library = librarySchema
-  }
-}
-
-/**
- * A specification for a tokenized tag group.
- */
-export class GroupSpec extends SubstringSpec {
-  /**
-   * The child group specifications.
-   * @type {GroupSpec[]}
-   */
-  children
-
-  constructor(start, end, children) {
-    super(start, end)
-
-    this.children = children
-  }
-}
-
-/**
- * A specification for a tokenized column splice template.
- */
-export class ColumnSpliceSpec extends SubstringSpec {
-  /**
-   * The column name this spec refers to.
-   * @type {string}
-   */
-  columnName
-
-  constructor(name, start, end) {
-    super(start, end)
-
-    this.columnName = name.trim()
-  }
-}
-
-class TokenizerState {
-  constructor() {
-    this.currentToken = '' // Characters in the token currently being parsed
-    this.groupDepth = 0
-    this.startingIndex = 0 // Starting index of this token
-    this.lastDelimiter = [undefined, -1] // Type and position of the last delimiter
-    this.librarySchema = ''
-    this.columnSpliceIndex = -1 //Index of the last { if this token is column splice
-    this.currentGroupStack = [[]]
-    this.parenthesesStack = []
-  }
-}
-
-/**
- * Class for tokenizing HED strings.
- */
-export class HedStringTokenizerNew {
-  constructor(hedString) {
-    this.hedString = hedString
-    this.issues = []
-    this.state = null
-  }
-
-  /**
-   * Split the HED string into delimiters and tags.
-   *
-   * @returns {[TagSpec[], GroupSpec, Object<string, Issue[]>]} The tag specifications, group bounds, and any issues found.
-   */
-  tokenize() {
-    this.initializeTokenizer()
-
-    for (let i = 0; i < this.hedString.length; i++) {
-      const character = this.hedString.charAt(i)
-      this.handleCharacter(i, character)
-      //this.tokenizeCharacter(i, character)
-      if (this.issues.length > 0) {
-        return [null, null, this.issues]
-      }
-    }
-    this.finalizeTokenizer()
-    if (this.issues.length > 0) {
-      return [null, null, this.issues]
-    } else {
-      return [this.state.currentGroupStack.pop(), this.state.parenthesesStack.pop(), []]
-    }
-  }
-
-  resetToken(i) {
-    this.state.startingIndex = i + 1
-    this.state.currentToken = ''
-    this.state.librarySchema = ''
-  }
-
-  finalizeTokenizer() {
-    if (this.state.lastDelimiter[0] === CHARACTERS.OPENING_COLUMN) {
-      // Extra opening brace
-      this.pushIssue('unclosedCurlyBrace', this.state.lastDelimiter[1])
-    } else if (this.state.lastDelimiter[0] === CHARACTERS.OPENING_GROUP) {
-      // Extra opening parenthesis
-      this.pushIssue('unclosedParentheses', this.state.lastDelimiter[1])
-    } else if (
-      this.state.lastDelimiter[0] === CHARACTERS.COMMA &&
-      this.hedString.slice(this.state.lastDelimiter[1] + 1).trim().length === 0
-    ) {
-      this.pushIssue('emptyTagFound', this.state.lastDelimiter[1]) // Extra comma
-    } else {
-      this.unwindGroupStack()
-    }
-  }
-
-  initializeTokenizer() {
-    this.issues = []
-    this.state = new TokenizerState()
-    this.state.parenthesesStack = [new GroupSpec(0, this.hedString.length, [])]
-  }
-
-  handleCharacter(i, character) {
-    const characterHandler = {
-      [CHARACTERS.OPENING_GROUP]: () => this.handleOpeningGroup(i),
-      [CHARACTERS.CLOSING_GROUP]: () => this.handleClosingGroup(i),
-      [CHARACTERS.OPENING_COLUMN]: () => this.handleOpeningColumn(i),
-      [CHARACTERS.CLOSING_COLUMN]: () => this.handleClosingColumn(i),
-      [CHARACTERS.COMMA]: () => this.handleComma(i),
-      [CHARACTERS.COLON]: () => this.handleColon(character),
-    }[character] // Selects the character handler based on the value of character
-
-    if (characterHandler) {
-      characterHandler()
-    } else if (invalidCharacters.has(character)) {
-      this.pushInvalidCharacterIssue(character, i)
-    } else {
-      this.state.currentToken += character
-    }
-  }
-
-  handleComma(i) {
-    if (
-      this.state.lastDelimiter[0] === CHARACTERS.COMMA &&
-      this.hedString.slice(this.state.lastDelimiter[1] + 1, i).trim().length === 0
-    ) {
-      this.pushIssue('emptyTagFound', this.state.lastDelimiter[1]) // Check for empty group between commas
-    } else if (
-      this.state.currentToken.trim().length === 0 &&
-      [CHARACTERS.CLOSING_GROUP, CHARACTERS.CLOSING_COLUMN].includes(this.state.lastDelimiter[0])
-    ) {
-      this.resetToken(i)
-    } else {
-      this.pushTag(i)
-      this.state.lastDelimiter = [CHARACTERS.COMMA, i]
-    }
-  }
-
-  handleOpeningGroup(i) {
-    if (this.state.lastDelimiter[0] === CHARACTERS.OPENING_COLUMN) {
-      this.pushIssue('unclosedCurlyBrace', this.state.lastDelimiter[1])
-    } else {
-      this.state.currentGroupStack.push([])
-      this.state.parenthesesStack.push(new GroupSpec(i, undefined, []))
-      this.resetToken(i)
-      this.state.groupDepth++
-      this.state.lastDelimiter = [CHARACTERS.OPENING_GROUP, i]
-    }
-  }
-
-  handleClosingGroup(i) {
-    if (this.state.currentToken.trim().length > 0) {
-      // only push a tag if it has length > 0. Empty groups are allowed.
-      this.pushTag(i)
-    }
-    if (this.state.groupDepth <= 0) {
-      // If the group depth is <= 0, it means there's no corresponding opening group.
-      this.pushIssue('unopenedParenthesis', i)
-    } else if (this.state.lastDelimiter[0] === CHARACTERS.OPENING_COLUMN) {
-      this.pushIssue('unclosedCurlyBrace', this.state.lastDelimiter[1])
-    } else {
-      // Close the group by updating its bounds and moving it to the parent group.
-      this.closeGroup(i)
-      this.state.lastDelimiter = [CHARACTERS.CLOSING_GROUP, i]
-    }
-  }
-
-  handleOpeningColumn(i) {
-    if (this.state.currentToken.trim().length > 0) {
-      // In the middle of a token -- can't have an opening brace
-      this.pushInvalidCharacterIssue(CHARACTERS.OPENING_COLUMN, i)
-    } else if (this.state.lastDelimiter[0] === CHARACTERS.OPENING_COLUMN) {
-      //
-      this.pushIssue('nestedCurlyBrace', i)
-    } else {
-      this.state.lastDelimiter = [CHARACTERS.OPENING_COLUMN, i]
-    }
-  }
-
-  handleClosingColumn(i) {
-    if (this.state.lastDelimiter !== CHARACTERS.OPENING_COLUMN) {
-      // Column splice not in progress
-      this.pushIssue('unopenedCurlyBrace', i)
-    } else if (!this.state.currentToken.trim()) {
-      // Ensure that column slice is not empty
-      this.pushIssue('emptyCurlyBrace', i)
-    } else {
-      // Close column by updating bounds and moving it to the parent group, push a column splice on the stack.
-      this.state.currentGroupStack[this.state.groupDepth].push(
-        new ColumnSpliceSpec(this.state.currentToken.trim(), this.state.startingIndex, i),
-      )
-      this.resetToken(i)
-      this.state.lastDelimiter = [CHARACTERS.CLOSING_COLUMN, i]
-    }
-  }
-
-  handleColon(i) {
-    if (!this.state.librarySchema) {
-      // If a colon has not been seen -- its a library. Ignore other colons.
-      this.state.librarySchema = this.state.currentToken
-      this.resetToken(i)
-    } else {
-      this.state.currentToken += CHARACTERS.COLON
-    }
-  }
-
-  unwindGroupStack() {
-    while (this.state.groupDepth > 0) {
-      this.pushIssue(
-        'unclosedParenthesis',
-        this.state.parenthesesStack[this.state.parenthesesStack.length - 1].bounds[0],
-      )
-      this.closeGroup(this.hedString.length)
-    }
-  }
-
-  pushTag(i) {
-    // Called when a token has been parsed
-    const token = this.state.currentToken.trim()
-    if (!token) {
-      // Empty tokens cannot be pushed
-      this.pushIssue('emptyTagFound', i)
-    } else if (token.startsWith(CHARACTERS.SLASH) || token.endsWith(CHARACTERS.SLASH)) {
-      this.pushIssue('extraSlash', this.state.startingIndex)
-    } else {
-      const bounds = getTrimmedBounds(this.state.currentToken)
-      this.state.currentGroupStack[this.state.groupDepth].push(
-        new TagSpec(
-          this.state.currentToken.trim(),
-          this.state.startingIndex + bounds[0],
-          this.state.startingIndex + bounds[1],
-          this.state.librarySchema,
-        ),
-      )
-      this.resetToken(i)
-    }
-  }
-
-  closeGroup(i) {
-    const groupSpec = this.state.parenthesesStack.pop()
-    groupSpec.bounds[1] = i + 1
-    this.state.parenthesesStack[this.state.groupDepth - 1].children.push(groupSpec)
-    this.state.currentGroupStack[this.state.groupDepth - 1].push(this.state.currentGroupStack.pop())
-    this.state.groupDepth--
-    this.resetToken(i)
-  }
-
-  checkValueTagForInvalidCharacters() {
-    const formToCheck = replaceTagNameWithPound(this.state.currentToken)
-    for (let i = 0; i < formToCheck.length; i++) {
-      const character = formToCheck.charAt(i)
-      if (invalidCharactersOutsideOfValues.has(character)) {
-        this.pushInvalidCharacterIssue(character, this.state.startingIndex + i)
-      }
-    }
-  }
-
-  pushIssue(issueCode, index) {
-    this.issues.push(generateIssue(issueCode, { index, string: this.hedString }))
-  }
-
-  pushInvalidCharacterIssue(character, index) {
-    this.issues.push(
-      generateIssue('invalidCharacter', { character: unicodeName(character), index, string: this.hedString }),
-    )
-  }
-}
